// src/app/shared/models/device-info.model.ts
export interface DeviceInfo {
    AppName: string;
    AppVersionCode: string;
    AppVersionName: string;
    DeviceName: string;
    IMEI: string;
    MAC: string;
    Model: string;
    PhoneNumber: string;
    Platform: number;
    Port: number;
    PrivateIP: string;
    Resolution: string;
    SDKVersionName: string;
    SerialNum: string;
}